- redesigned the user registration and authentication workflow based on OAuth2 best practices - simplified the creation and update of a resource.

+++ fundraiser-service/src/main/java/com/gohelpfund/api/v1/fundraisers/controllers/FundraiserController.java
@@ -10,10 +10,8 @@ import org.springframework.hateoas.ResourceSupport;
 import org.springframework.hateoas.Resources;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -46,40 +44,53 @@ public class FundraiserController {
     @GetMapping("/{fundraiserId}")
     public Resource<Fundraiser> one(@PathVariable("fundraiserId") String fundraiserId) {
 
-        return assembler.toResource(
-                service.getOne(fundraiserId)
-                        .orElseThrow(() -> new EntityNotFoundException(Fundraiser.class, "id", fundraiserId)));
+        Fundraiser fundraiser = service.getOne(fundraiserId);
+
+        if(fundraiser == null){
+            throw new EntityNotFoundException(Fundraiser.class, "id", fundraiserId);
+        }
+        return assembler.toResource(fundraiser);
+
     }
 
     @PostMapping()
-    public ResponseEntity<Resource<Fundraiser>> setFundraiser(@RequestBody @Valid Fundraiser fundraiser) {
-        Fundraiser newFundraiser = service.save(fundraiser);
+    public ResponseEntity<Resource<Fundraiser>> setFundraiser() {
+        Fundraiser newFundraiser = service.save();
         return ResponseEntity
                 .created(linkTo(methodOn(FundraiserController.class).one(newFundraiser.getFundraiserId())).toUri())
                 .body(assembler.toResource(newFundraiser));
     }
 
-    @DeleteMapping("/{fundraiserId}")
+    @PutMapping("/{fundraiserId}")
+    public ResponseEntity<ResourceSupport> complete(@PathVariable String fundraiserId) {
+
+        Fundraiser fundraiser = service.getOne(fundraiserId);
+
+        if(fundraiser == null){
+            throw new EntityNotFoundException(Fundraiser.class, "id", fundraiserId);
+        }
+
+        // TODO: 15-Apr-19  implement update logic of a fundraiser
+        service.save();
+
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(null);
+    }
+
+/*    @DeleteMapping("/{fundraiserId}")
     public ResponseEntity<ResourceSupport> cancel(@PathVariable String fundraiserId) {
-        Fundraiser fundraiser = service.getOne(fundraiserId).orElseThrow(() -> new EntityNotFoundException(Fundraiser.class, "id", fundraiserId));
+        Fundraiser fundraiser = service.getOne(fundraiserId);
+
+        if(fundraiser == null){
+            throw new EntityNotFoundException(Fundraiser.class, "id", fundraiserId);
+        }
 
         service.delete(fundraiser.getFundraiserId());
 
         return ResponseEntity
                 .status(HttpStatus.NO_CONTENT)
                 .body(null);
-    }
-
-    @PutMapping("/{fundraiserId}")
-    public ResponseEntity<ResourceSupport> complete(@PathVariable String fundraiserId) {
-
-        Fundraiser fundraiser = service.getOne(fundraiserId).orElseThrow(() -> new EntityNotFoundException(Fundraiser.class, "id", fundraiserId));
-
-        service.save(fundraiser);
-
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(null);
-    }
+    }*/
 
 }
